requirements.txt
Flask
requests
openai
psycopg2-binary